requirements.txt
streamlit==1.28.1
ultralytics==8.0.196
opencv-python-headless==4.8.1.78
Pillow==10.0.1
numpy==1.24.3
pandas==2.0.3
psycopg2-binary==2.9.7
sqlalchemy==2.0.21
mlflow==2.8.1
torch==2.0.1
torchvision==0.15.2
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0
python-multipart==0.0.6